Move the resource to it's own package
