+++ internal/elasticsearch/security/system_user/resource.go
@@ -1,0 +1,26 @@
+package system_user
+
+import (
+	"context"
+
+	"github.com/elastic/terraform-provider-elasticstack/internal/clients"
+	"github.com/hashicorp/terraform-plugin-framework/resource"
+)
+
+func NewSystemUserResource() resource.Resource {
+	return &systemUserResource{}
+}
+
+type systemUserResource struct {
+	client *clients.ApiClient
+}
+
+func (r *systemUserResource) Metadata(_ context.Context, req resource.MetadataRequest, resp *resource.MetadataResponse) {
+	resp.TypeName = req.ProviderTypeName + "_elasticsearch_security_system_user"
+}
+
+func (r *systemUserResource) Configure(_ context.Context, req resource.ConfigureRequest, resp *resource.ConfigureResponse) {
+	client, diags := clients.ConvertProviderData(req.ProviderData)
+	resp.Diagnostics.Append(diags...)
+	r.client = client
+}
